+++ src/RequestInterface.php
@@ -8,6 +8,11 @@ namespace Speicher210\Monsum\Api;
 interface RequestInterface
 {
     /**
+     * Maximum number of elements for the retrieval defined in Monsum.
+     */
+    const MAX_LIST_LIMIT = 100;
+
+    /**
      * Get the name of the service.
      *
      * @return string